# backend-api-deprecated/docs/api/versions/1.0.0/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "实验室管理系统后端 API",
    "version": "1.0.0",
    "description": "\n实验室管理系统后端 API 是一个 RESTful API 服务，为实验室智能管理系统前端提供数据存储、业务逻辑处理和系统集成能力。\n\n## 核心功能\n\n- **样品管理**：样品全生命周期管理、流转追踪、分样合样\n- **工作流引擎**：灵活的工作流配置、任务自动派工\n- **检测结果**：结果录入、公式计算、批量导入、异常检测\n- **审核判定**：多级审核流程、质量判定、样品放行\n- **报告管理**：报告模板、报告生成、电子签名、分发回收\n- **统计分析**：数据聚合、自定义报表、数据导出\n- **系统管理**：用户管理、角色权限、审计日志、数据备份\n\n## 认证方式\n\nAPI 使用 JWT (JSON Web Token) 进行身份认证。\n\n1. 调用 `POST /api/auth/login` 获取访问令牌\n2. 在后续请求的 Header 中添加：`Authorization: Bearer <token>`\n3. 访问令牌有效期为 15 分钟，可使用刷新令牌延长会话\n\n## 错误响应格式\n\n所有错误响应遵循统一格式：\n\n```json\n{\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"用户友好的错误消息\",\n    \"details\": {},\n    \"timestamp\": \"2024-01-01T00:00:00.000Z\",\n    \"path\": \"/api/samples\",\n    \"requestId\": \"uuid\"\n  }\n}\n```\n\n## 分页响应格式\n\n列表查询接口支持分页，响应格式：\n\n```json\n{\n  \"items\": [],\n  \"total\": 100,\n  \"page\": 1,\n  \"pageSize\": 20,\n  \"totalPages\": 5\n}\n```\n      ",
    "contact": {
      "name": "API 支持",
      "email": "support@lims.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-build-info": {
      "buildTime": "2026-03-10T02:44:28.444Z",
      "environment": "local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 访问令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "错误码",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "description": "错误消息",
                "example": "请求参数验证失败"
              },
              "details": {
                "type": "object",
                "description": "详细错误信息"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time",
                "description": "错误发生时间"
              },
              "path": {
                "type": "string",
                "description": "请求路径"
              },
              "requestId": {
                "type": "string",
                "description": "请求追踪 ID"
              }
            }
          }
        }
      },
...
      "Sample": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "样品 ID"
          },
          "barcode": {
            "type": "string",
            "description": "样品条码（系统自动生成）",
            "example": "SAMPLE-20240101-001"
          },
          "sampleNumber": {
            "type": "string",
            "description": "样品编号（系统自动生成）",
            "example": "S20240101001"
          },
          "clientName": {
            "type": "string",
            "description": "客户名称",
            "example": "测试公司"
          },
          "clientContact": {
            "type": "string",
            "description": "客户联系方式",
            "example": "张三 13800138000"
          },
          "sampleName": {
            "type": "string",
            "description": "样品名称",
            "example": "饮用水样品"
          },
          "sampleType": {
            "type": "string",
            "description": "样品类型",
            "example": "水质"
          },
          "sampleCategory": {
            "type": "string",
            "description": "样品类别",
            "example": "环境样品"
          },
          "quantity": {
            "type": "number",
            "format": "float",
            "description": "样品数量",
            "example": 500
          },
          "unit": {
            "type": "string",
            "description": "数量单位",
            "example": "mL"
          },
          "receivedDate": {
            "type": "string",
            "format": "date-time",
            "description": "接收日期"
          },
          "samplingDate": {
            "type": "string",
            "format": "date-time",
            "description": "采样日期"
          },
          "samplingLocation": {
            "type": "string",
            "description": "采样地点",
            "example": "某市自来水厂"
          },
          "samplingPerson": {
            "type": "string",
            "description": "采样人员",
            "example": "李四"
          },
          "storageLocation": {
            "type": "string",
            "description": "存储位置",
            "example": "冷藏室-A区-01号柜"
          },
          "storageCondition": {
            "type": "string",
            "description": "存储条件",
            "example": "4℃冷藏"
          },
          "status": {
            "type": "string",
            "enum": [
              "REGISTERED",
              "IN_TESTING",
              "TESTING_COMPLETE",
              "IN_AUDIT",
              "AUDIT_COMPLETE",
              "RELEASED",
              "ARCHIVED"
            ],
            "description": "样品状态",
            "example": "REGISTERED"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH",
              "URGENT"
            ],
            "description": "优先级",
            "example": "NORMAL"
          },
          "description": {
            "type": "string",
            "description": "样品描述"
          },
          "remarks": {
            "type": "string",
            "description": "备注信息"
          },
          "parentSampleId": {
            "type": "string",
            "format": "uuid",
            "description": "母样品 ID（分样时使用）"
          },
          "mergedFromIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "合样来源样品 ID 列表"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid",
            "description": "创建人 ID"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time",
            "description": "放行时间"
          },
          "releasedBy": {
            "type": "string",
            "format": "uuid",
            "description": "放行人 ID"
          }
        }
      },
...
      "Transfer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "流转记录 ID"
          },
          "sampleId": {
            "type": "string",
            "format": "uuid",
            "description": "样品 ID"
          },
          "fromLocation": {
            "type": "string",
            "description": "起始位置",
            "example": "接收室"
          },
          "toLocation": {
            "type": "string",
            "description": "目标位置",
            "example": "检测室-A"
          },
          "fromPerson": {
            "type": "string",
            "description": "交接人",
            "example": "张三"
          },
          "toPerson": {
            "type": "string",
            "description": "接收人",
            "example": "李四"
          },
          "transferDate": {
            "type": "string",
            "format": "date-time",
            "description": "流转日期"
          },
          "receivedDate": {
            "type": "string",
            "format": "date-time",
            "description": "接收确认日期"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_TRANSIT",
              "RECEIVED",
              "REJECTED"
            ],
            "description": "流转状态"
          },
          "remarks": {
            "type": "string",
            "description": "备注"
          },
          "senderConfirmed": {
            "type": "boolean",
            "description": "发送方是否已确认"
          },
          "receiverConfirmed": {
            "type": "boolean",
            "description": "接收方是否已确认"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "未认证或令牌无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "AUTH_FAILED",
                "message": "认证失败，请重新登录",
                "details": {
                  "reason": "令牌已过期"
                },
                "timestamp": "2024-01-01T00:00:00.000Z",
                "path": "/api/samples",
                "requestId": "uuid"
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "无权限访问",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "PERMISSION_DENIED",
                "message": "您没有权限执行此操作",
                "details": {
                  "required": "sample:update",
                  "current": [
                    "sample:read"
                  ]
                },
                "timestamp": "2024-01-01T00:00:00.000Z",
                "path": "/api/samples/123",
                "requestId": "uuid"
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "请求的资源不存在",
                "timestamp": "2024-01-01T00:00:00.000Z",
                "path": "/api/samples/123",
                "requestId": "uuid"
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "请求参数验证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "请求参数验证失败",
                "details": {
                  "fields": [
                    {
                      "field": "email",
                      "message": "邮箱格式不正确"
                    },
                    {
                      "field": "quantity",
                      "message": "数量必须大于 0"
                    }
                  ]
                },
                "timestamp": "2024-01-01T00:00:00.000Z",
                "path": "/api/samples",
                "requestId": "uuid"
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "服务器内部错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "服务器内部错误，请稍后重试",
                "timestamp": "2024-01-01T00:00:00.000Z",
                "path": "/api/samples",
                "requestId": "uuid"
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/samples/{id}/transfer": {
      "post": {
        "summary": "样品流转",
        "description": "创建样品流转记录，更新样品当前位置，支持双方确认机制",
        "tags": [
          "样品管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "样品 ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fromLocation",
                  "toLocation",
                  "fromPerson",
                  "toPerson"
                ],
                "properties": {
                  "fromLocation": {
                    "type": "string",
                    "description": "起始位置",
                    "example": "接收室"
                  },
                  "toLocation": {
                    "type": "string",
                    "description": "目标位置",
                    "example": "检测室-A"
                  },
                  "fromPerson": {
                    "type": "string",
                    "description": "交接人",
                    "example": "张三"
                  },
                  "toPerson": {
                    "type": "string",
                    "description": "接收人",
                    "example": "李四"
                  },
                  "remarks": {
                    "type": "string",
                    "description": "备注信息"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "流转成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "样品流转成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Transfer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/samples/{id}/split": {
      "post": {
        "summary": "分样操作",
        "description": "从母样品创建多个子样品，建立母子关联关系，操作在事务中执行",
        "tags": [
          "样品管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "母样品 ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "childSamples"
                ],
                "properties": {
                  "childSamples": {
                    "type": "array",
                    "description": "子样品信息列表",
                    "items": {
                      "type": "object",
                      "required": [
                        "sampleName",
                        "quantity",
                        "unit"
                      ],
                      "properties": {
                        "sampleName": {
                          "type": "string",
                          "description": "子样品名称",
                          "example": "水样-1"
                        },
                        "quantity": {
                          "type": "number",
                          "description": "数量",
                          "example": 100
                        },
                        "unit": {
                          "type": "string",
                          "description": "单位",
                          "example": "mL"
                        },
                        "storageLocation": {
                          "type": "string",
                          "description": "存储位置"
                        },
                        "remarks": {
                          "type": "string",
                          "description": "备注"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "分样成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "分样成功"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Sample"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/samples/merge": {
      "post": {
        "summary": "合样操作",
        "description": "将多个样品合并为一个新样品，记录来源样品信息，操作在事务中执行",
        "tags": [
          "样品管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sourceSampleIds",
                  "mergedSample"
                ],
                "properties": {
                  "sourceSampleIds": {
                    "type": "array",
                    "description": "来源样品 ID 列表",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "example": [
                      "uuid1",
                      "uuid2",
                      "uuid3"
                    ]
                  },
                  "mergedSample": {
                    "type": "object",
                    "description": "合并后的样品信息",
                    "required": [
                      "sampleName",
                      "quantity",
                      "unit"
                    ],
                    "properties": {
                      "sampleName": {
                        "type": "string",
                        "description": "合并样品名称",
                        "example": "混合水样"
                      },
                      "quantity": {
                        "type": "number",
                        "description": "数量",
                        "example": 500
                      },
                      "unit": {
                        "type": "string",
                        "description": "单位",
                        "example": "mL"
                      },
                      "storageLocation": {
                        "type": "string",
                        "description": "存储位置"
                      },
                      "remarks": {
                        "type": "string",
                        "description": "备注"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "合样成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "合样成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Sample"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}